+++ backend/gb/src/main/java/com/gamebox/gb/services/WishlistService.java
@@ -8,7 +8,10 @@ import com.gamebox.gb.domain.dtos.wishlist.WishlistResponse;
 import com.gamebox.gb.domain.dtos.wishlist.WishlistSearchResponse;
 import com.gamebox.gb.domain.entities.Game;
 import com.gamebox.gb.domain.entities.Profile;
+import com.gamebox.gb.domain.entities.User;
 import com.gamebox.gb.domain.entities.Wishlist;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -32,6 +35,14 @@ public class WishlistService {
 
         Profile profile = profileRepository.findById(request.profileId())
                 .orElseThrow(() -> new RuntimeException("Perfil não encontrado."));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        User loggedUser = (User) authentication.getPrincipal();
+
+        if(!profile.getUser().getId().equals(loggedUser.getId())) {
+            throw new RuntimeException("Acesso negado.");
+        }
 
         Game game = gameRepository.findById(request.gameId())
                 .orElseThrow(() -> new RuntimeException("Jogo não encontrado."));
@@ -75,6 +86,14 @@ public class WishlistService {
 
         Wishlist wishlist = wishlistRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Registro não encontrado."));
+
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        User loggedUser = (User) authentication.getPrincipal();
+
+        if (!wishlist.getProfile().getUser().getId().equals(loggedUser.getId())) {
+            throw new RuntimeException("Acesso negado.");
+        }
 
         wishlistRepository.delete(wishlist);
     }
